+++ LetsDoIt.Moody/src/LetsDoIt.Moody.Web/Entities/Requests/CategoryInsertRequest.cs
@@ -8,7 +8,7 @@ namespace LetsDoIt.Moody.Web.Entities.Requests
         [Required]
         public string Name { get; set; }
 
-        public int Order { get; set; }
+        public decimal Order { get; set; }
 
         [Required]
         public string Image { get; set; }
